requirements.txt
asgiref==3.2.7
certifi==2018.10.15
Django==3.0.4
django-appconf==1.0.2
django-axes==4.4.0
django-cors-middleware==1.3.1
django-filter==2.1.0
django-ipware==2.1.0
djangorestframework==3.11.0
numpy==1.18.2
pandas==1.0.3
Pillow==5.3.0
psycopg2-binary==2.7.6
python-dateutil==2.8.1
pytz==2018.5
six==1.14.0
sqlparse==0.3.1
